requirements.txt
fastapi
python-dotenv
langchain
langchain-core
langchain-community
uvicorn
langchainhub
faiss-cpu
langchain-fireworks
langchain-cohere
poetry